# docs/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: CMS New Community API
  version: 0.1.0
  description: REST API contract for a Korean large-community web service.
servers:
  - url: https://api.example.com
    description: Production placeholder
  - url: http://localhost:8080
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/boards:
    get:
      tags: [Boards]
      summary: List visible boards
      security: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Board list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedBoards'
    post:
      tags: [Admin]
      summary: Create a board
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BoardUpsertRequest'
      responses:
        '201':
          description: Board created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Board'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/boards/{boardId}:
    get:
      tags: [Boards]
      summary: Get board detail
      security: []
      parameters:
        - $ref: '#/components/parameters/BoardId'
      responses:
        '200':
          description: Board detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Board'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Admin]
      summary: Update board settings
      parameters:
        - $ref: '#/components/parameters/BoardId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BoardUpsertRequest'
      responses:
        '200':
          description: Board updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Board'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/boards/{boardId}/subscription:
    put:
      tags: [Boards]
      summary: Subscribe to a board
      parameters:
        - $ref: '#/components/parameters/BoardId'
      responses:
        '204':
          description: Subscribed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Boards]
      summary: Unsubscribe from a board
      parameters:
        - $ref: '#/components/parameters/BoardId'
      responses:
        '204':
          description: Unsubscribed
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/posts/{postId}/comments:
    get:
      tags: [Comments]
      summary: List comments for a post
      security: []
      parameters:
        - $ref: '#/components/parameters/PostId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Comment list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedComments'
    post:
      tags: [Comments]
      summary: Create a comment
      parameters:
        - $ref: '#/components/parameters/PostId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentCreateRequest'
      responses:
        '201':
          description: Comment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/notifications:
    get:
      tags: [Notifications]
      summary: List current member notifications
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
            default: false
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Notification list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedNotifications'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/notifications/{notificationId}/read:
    patch:
      tags: [Notifications]
      summary: Mark a notification as read
      parameters:
        - $ref: '#/components/parameters/NotificationId'
      responses:
        '204':
          description: Notification marked read
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/reports:
    post:
      tags: [Reports]
      summary: Report a post or comment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportCreateRequest'
      responses:
        '201':
          description: Report submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/admin/reports:
    get:
      tags: [Admin]
      summary: List reports for moderation
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [OPEN, IN_REVIEW, ACTIONED, REJECTED, DUPLICATE]
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Report queue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedReports'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/reports/{reportId}/actions:
    post:
      tags: [Admin]
      summary: Record a moderation action for a report
      parameters:
        - $ref: '#/components/parameters/ReportId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModerationActionRequest'
      responses:
        '201':
          description: Action recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModerationAction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    BoardId:
      name: boardId
      in: path
      required: true
      schema:
        type: string
    PostId:
      name: postId
      in: path
      required: true
      schema:
        type: string
    NotificationId:
      name: notificationId
      in: path
      required: true
      schema:
        type: string
    ReportId:
      name: reportId
      in: path
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Not allowed for the current member
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    Board:
      type: object
      required: [id, slug, name, visibility, postingPolicy]
      properties:
        id:
          type: string
        slug:
          type: string
        name:
          type: string
        description:
          type: string
        visibility:
          type: string
          enum: [PUBLIC, MEMBERS_ONLY, CLOSED]
        postingPolicy:
          type: string
          enum: [MEMBERS, MODERATORS, ADMINS]
        categoryOptions:
          type: array
          items:
            type: string
        sortOrder:
          type: integer
        isArchived:
          type: boolean
    BoardUpsertRequest:
      type: object
      required: [slug, name, visibility, postingPolicy]
      properties:
        slug:
          type: string
        name:
          type: string
        description:
          type: string
        visibility:
          type: string
          enum: [PUBLIC, MEMBERS_ONLY, CLOSED]
        postingPolicy:
          type: string
          enum: [MEMBERS, MODERATORS, ADMINS]
        categoryOptions:
          type: array
          items:
            type: string
        sortOrder:
          type: integer
        isArchived:
          type: boolean
...
    Comment:
      type: object
      required: [id, postId, authorNickname, body, status, createdAt]
      properties:
        id:
          type: string
        postId:
          type: string
        parentCommentId:
          type: string
          nullable: true
        authorNickname:
          type: string
        body:
          type: string
        status:
          type: string
          enum: [PUBLISHED, HIDDEN, DELETED]
        depth:
          type: integer
        createdAt:
          type: string
          format: date-time
    CommentCreateRequest:
      type: object
      required: [body]
      properties:
        parentCommentId:
          type: string
          nullable: true
        body:
          type: string
          minLength: 1
          maxLength: 2000
...
    Notification:
      type: object
      required: [id, type, title, message, createdAt]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [REPLY, COMMENT_ON_POST, MODERATION_RESULT, NOTICE]
        title:
          type: string
        message:
          type: string
        targetType:
          type: string
          nullable: true
        targetId:
          type: string
          nullable: true
        readAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    ReportCreateRequest:
      type: object
      required: [targetType, targetId, reasonCode]
      properties:
        targetType:
          type: string
          enum: [POST, COMMENT]
        targetId:
          type: string
        reasonCode:
          type: string
          enum: [SPAM, ABUSE, ILLEGAL, PRIVACY, OTHER]
        description:
          type: string
          maxLength: 1000
    Report:
      type: object
      required: [id, targetType, targetId, reasonCode, status, createdAt]
      properties:
        id:
          type: string
        reporterNickname:
          type: string
        targetType:
          type: string
          enum: [POST, COMMENT]
        targetId:
          type: string
        reasonCode:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [OPEN, IN_REVIEW, ACTIONED, REJECTED, DUPLICATE]
        createdAt:
          type: string
          format: date-time
    ModerationActionRequest:
      type: object
      required: [actionType, reason]
      properties:
        actionType:
          type: string
          enum: [HIDE, RESTORE, DELETE, WARN, SUSPEND, REJECT_REPORT, MARK_DUPLICATE]
        reason:
          type: string
          minLength: 3
        visibleToMember:
          type: boolean
          default: true
    ModerationAction:
      type: object
      required: [id, actorNickname, actionType, reason, createdAt]
      properties:
        id:
          type: string
        actorNickname:
          type: string
        reportId:
          type: string
          nullable: true
        actionType:
          type: string
        reason:
          type: string
        visibleToMember:
          type: boolean
        createdAt:
          type: string
          format: date-time
    PagedBoards:
      allOf:
        - $ref: '#/components/schemas/PageEnvelope'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/Board'
...
    PagedComments:
      allOf:
        - $ref: '#/components/schemas/PageEnvelope'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/Comment'
    PagedNotifications:
      allOf:
        - $ref: '#/components/schemas/PageEnvelope'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/Notification'
    PagedReports:
      allOf:
        - $ref: '#/components/schemas/PageEnvelope'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/Report'
    PageEnvelope:
      type: object
      required: [page, size, totalElements, totalPages]
      properties:
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        fieldErrors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string